requirements.txt
torch
numpy
scipy
matplotlib
seaborn
pandas
tqdm
pyyaml